Standardize fill_window implementations and abstract out slide_hash_neon for ARM.
