+++ arch/x86/README.md
@@ -3,6 +3,6 @@ Contents
 
 |Name|Description|
 |:-|:-|
-|fill_window_sse.c|SSE2 optimized fill_window|
 |deflate_quick.c|SSE4 optimized deflate strategy for use as level 1|
 |crc_folding.c|SSE4 + PCLMULQDQ optimized CRC folding implementation|
+|slide_sse2.c|SSE2 optimized slide_hash|
