+++ fallback_builtins.h
@@ -12,6 +12,7 @@
  */
 static __forceinline unsigned long __builtin_ctzl(unsigned long value) {
 #ifdef X86_CPUID
+    x86_check_features();
     if (x86_cpu_has_tzcnt)
         return _tzcnt_u32(value);
 #endif
