+++ functable.h
@@ -9,7 +9,6 @@
 #include "deflate.h"
 
 struct functable_s {
-    void     (* fill_window)        (deflate_state *s);
     Pos      (* insert_string)      (deflate_state *const s, const Pos str, unsigned int count);
     Pos      (* quick_insert_string)(deflate_state *const s, const Pos str);
     uint32_t (* adler32)            (uint32_t adler, const unsigned char *buf, size_t len);
